requirements.txt
Django==5.2.2
djangorestframework==3.16.1
django-cors-headers==4.9.0
djangorestframework-simplejwt==5.5.1
gunicorn==21.2.0